Quite a big update!!!
FIxed a lot in this push, Added a renewed login system which is more OOP based. Also got the whole application to know which user is logged in. Added a greeting to the home page with the logged in user an d a dynamic greeting based on the users time. Also fixed that the user can press enter to log in and the password is now hidden with dots. AKA i have a headache